Documentation Index

Here's an index of the available functions.

Quantum_Geometric_Complexity.:⊗

Quantum_Geometric_Complexity.CS_basis_bosons

Quantum_Geometric_Complexity.Clock

Quantum_Geometric_Complexity.IPR

Quantum_Geometric_Complexity.M

Quantum_Geometric_Complexity.purity

Quantum_Geometric_Complexity.tensor_list

Quantum_Geometric_Complexity.ρ_mat

Quantum_Geometric_Complexity.ρ_vec

Quantum_Geometric_Complexity.σ_n

Documentation

Quantum_Geometric_Complexity.:⊗ — Method

⊗(A, B)

Computes the Kronecker product of two matrices A and B.

Quantum_Geometric_Complexity.CS_basis_bosons — Method

Performance comment

Some benchmarks:

julia> @btime O_basis_bosons(20)
885.291 μs (1286 allocations: 11.59 MiB)
20×20×400 Array{ComplexF64, 3}:

julia> @btime O_basis_bosons(100)
1.105 s (51410 allocations: 6.78 GiB)

so we seem to be limited by memory unsurprisingly again.

Quantum_Geometric_Complexity.Clock — Method

Clock(N)

FINISH!!

Return the Clock matrix...

Performance comment

It is much more efficient if returned in Diagonal object: ```julia-repl julia> @btime Clock(1000)^2 # Diagonal object v 11.125 μs (3 allocations: 39.75 KiB) 1000×1000 Diagonal{ComplexF64, Vector{ComplexF64}}: (output ommitted)

julia> @btime Clock_diagm(1000)^2 # Matrix object 37.858 ms (6 allocations: 30.54 MiB) 1000×1000 Matrix{ComplexF64}: (output ommitted) ````

Quantum_Geometric_Complexity.IPR — Method

IPR(v)

Function for calculating the inverse participation ratio.

Mathematically, given some vector $v$ in some basis, this is defined as

$IPR(v) ≐ \sum_n || v_n ||^4.$

Examples

julia> IPR( [1/√(4), 1/√(4), 1/√(4), 1/√(4)]  )
0.25

julia> IPR( [1, 0, 0, 0]  )
1

Quantum_Geometric_Complexity.M — Method

M(ρ, B_basis)

Function for calculating the macroscopicity M of a density matrix in Hilbert space H×H.

Mathematically, given some matrix ρ in H×H, this is defined as

M ≐ -log2( IPR_C[ρ] / γ^2 ).

This is non trivial so please refer to the paper.

Examples

julia> M(I(2^3)/(2^3), B_basis_spin_half(3))
-0.0

julia> M(rand_ρ_pure(3), B_basis_spin_half(3))
4.062587703134164

Quantum_Geometric_Complexity.purity — Method

purity(ρ_μ::Vector)

Function for calculating the purity of a vector in characteristic space C.

Mathematically, given some vector ρ in characteristic space C, this is defined as

γ ≐ \sum_μ  ρ_μ ρ_μ^* ≡ ||ρ||^2

Examples

julia> ρ_vec_temp = ρ_vec(rand_ρ_pure(3), B_basis_spin_half(3))
64-element Vector{ComplexF64}:
0.35355339059327373 + 0.0im
-0.19777757955054376 - 4.336808689942018e-19im
-0.07369283640174634 + 5.204170427930421e-18im
0.030330980132771558 + 0.0im
0.052689065818701745 + 1.0408340855860843e-17im
                    ⋮
-0.01752572646095066 + 0.0im
-0.23462520253833954 + 0.0im
0.1193174939662679 - 4.336808689942018e-19im
0.10279530237835932 - 8.673617379884035e-18im
-0.10936868240286046 + 0.0im

julia> purity(ρ_vec_temp)
0.9999999999999999 + 0.0im

purity

julia> ρ_vec_temp = ρ_vec(I(2^3)/2^3, B_basis_spin_half(3))
64-element Vector{ComplexF64}:
0.35355339059327373 + 0.0im
                0.0 + 0.0im
                0.0 + 0.0im
                0.0 + 0.0im
                0.0 + 0.0im
                    ⋮
                0.0 + 0.0im
                0.0 + 0.0im
                0.0 + 0.0im
                0.0 + 0.0im
                0.0 + 0.0im

julia> purity(ρ_vec_temp)
0.12499999999999997

julia> 1/(2^3)
0.125

julia> ρ_vec_temp = ρ_vec(rand_ρ_pure(3), B_basis_spin_half(3))
64-element Vector{ComplexF64}:
0.35355339059327373 + 0.0im
-0.14065223381828953 + 5.204170427930421e-18im
0.10474649723834337 + 6.071532165918825e-18im
0.05213855626731899 + 0.0im
-0.23021903371920083 + 1.3010426069826053e-18im
                    ⋮
-0.007396885500585001 + 0.0im
-0.14963824202038384 + 0.0im
0.0038745649216279177 + 1.734723475976807e-18im
0.025673666375129044 + 6.071532165918825e-18im
0.022366636393981332 + 0.0im

julia> purity(ρ_vec_temp)
1.0

Quantum_Geometric_Complexity.tensor_list — Method

tensor_list(list_of_ops)

Computes the Kronecker product of a list of matrices.

Arguments

list_of_ops: A list of matrices.

Returns

The Kronecker product of all matrices in the list.

Quantum_Geometric_Complexity.ρ_mat — Method

ρ_mat(ρ, B_basis)

Inverse of ρ_mat, that is, converts vector in characteristic space $\mathcal{C}$ back to its Hilbert space representation in $\mathcal{H} \times \mathcal{H}$. That is, mathematically, it's a function

ρ_vec : C,O → H×H

Performance comment

This function also suffered from performance issues like ρ_vec(ρ, B_basis) which were also solved with Einsum.jl. For more details see ρ_vec(ρ, B_basis) documentation.

Examples

julia> B_basis = B_basis_spin_half(1)
2×2×4 Array{ComplexF64, 3}:
[:, :, 1] =
    0.707107+0.0im       0.0+0.0im
        0.0+0.0im  0.707107+0.0im

[:, :, 2] =
        0.0+0.0im  0.707107+0.0im
    0.707107+0.0im       0.0+0.0im

[:, :, 3] =
    0.0+0.0im       0.0+0.707107im
    0.0-0.707107im  0.0+0.0im

[:, :, 4] =
    0.707107+0.0im        0.0+0.0im
        0.0+0.0im  -0.707107+0.0im

julia> ψ = [1, 1]/√(2)
2-element Vector{Float64}:
    0.7071067811865475
    0.7071067811865475

julia> ρ_μ = ρ_vec(ψ*ψ', B_basis)
4-element Vector{ComplexF64}:
    0.7071067811865474 + 0.0im
    0.7071067811865474 + 0.0im
                0.0 + 0.0im
                0.0 + 0.0im

julia> ρ_mat(ρ_μ, B_basis)
2×2 Matrix{ComplexF64}:
    0.5+0.0im  0.5+0.0im
    0.5+0.0im  0.5+0.0im

julia> ψ*ψ'
2×2 Matrix{Float64}:
    0.5  0.5
    0.5  0.5

Quantum_Geometric_Complexity.ρ_vec — Method

ρ_vec(ρ, B_basis)

Converts density matrix $\rho \in \mathcal{H} \times \mathcal{H}$ to its vector representation in characteristic space $\mathcal{C}$ spanned by $\mathcal{B}$. That is, mathematically, it's a function

ρ_vec : H×H,B → C

Performance comment

This function used to be written as

function ρ_vec(ρ, B_basis)
    N = size(B_basis)[3]
    ρ_temp = ComplexF64.(zeros(N))
    for n=1:N
        ρ_temp[n] = tr(ρ*B_basis[:,:,n])
    end
    return ρ_temp
end

but it had performance issues:

julia> ρ = rand_ρ(7) # output ommited
julia> B_basis = B_basis_spin_half(7) # output ommited
julia> @btime ρ_vec(ρ,B_basis)
3.413 s (65540 allocations: 8.00 GiB)
(further output ommited)
julia> @btime ρ_vec_einsum(ρ,B_basis)
346.840 ms (2 allocations: 256.05 KiB)
(further output ommited)

so current default implementation is with Einsum.jl.

Examples

julia> B_basis = B_basis_spin_half(1)
2×2×4 Array{ComplexF64, 3}:
[:, :, 1] =
0.707107+0.0im       0.0+0.0im
    0.0+0.0im  0.707107+0.0im

[:, :, 2] =
    0.0+0.0im  0.707107+0.0im
0.707107+0.0im       0.0+0.0im

[:, :, 3] =
0.0+0.0im       0.0+0.707107im
0.0-0.707107im  0.0+0.0im

[:, :, 4] =
0.707107+0.0im        0.0+0.0im
    0.0+0.0im  -0.707107+0.0im

julia> ψ = [1, 1]/√(2)
2-element Vector{Float64}:
0.7071067811865475
0.7071067811865475

julia> ρ_vec(ψ*ψ', B_basis)
4-element Vector{ComplexF64}:
0.7071067811865474 + 0.0im
0.7071067811865474 + 0.0im
                0.0 + 0.0im
                0.0 + 0.0im

Quantum_Geometric_Complexity.σ_n — Method

σ_n(n)

Returns the Pauli matrix corresponding to the index n.

Arguments

n: Index of the Pauli matrix (0, 1, 2, 3 for I, X, Y, Z).

Returns

The corresponding Pauli matrix.